pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skill-miner"
version = "0.1.0"
description = "Mine reusable skills from your AI agent session history"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Andi Liao", email = "liaoandi95@gmail.com" }]
keywords = ["ai", "skills", "claude", "codex", "workflow", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
]

[project.scripts]
skill-miner = "skill_miner.cli:main"

[project.urls]
Homepage = "https://github.com/liaoandi/skill_miner"
Repository = "https://github.com/liaoandi/skill_miner"
Issues = "https://github.com/liaoandi/skill_miner/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/skill_miner"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
